requirements.txt
fastapi
uvicorn[standard]
requests
PyGithub
python-dotenv
google-generativeai